backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-extensions==3.2.3
requests==2.32.3
python-dotenv==1.0.0